requirements.txt
psutil==5.7.3
pyserial==3.4
